+++ src/StkTable/StkTable.vue
@@ -342,9 +342,10 @@ const props = withDefaults(
          */
         rowClassName?: (row: DT, i: number) => string;
         /**
-         * 列宽是否可拖动<br>
+         * 列宽是否可拖动(需要设置v-model:columns)<br>
          * **不要设置**列minWidth，**必须**设置width<br>
          * 列宽拖动时，每一列都必须要有width，且minWidth/maxWidth不生效。table width会变为"fit-content"。
+         * - 会自动更新props.columns中的with属性
          */
         colResizable?: boolean;
         /** 可拖动至最小的列宽 */
@@ -628,7 +629,7 @@ const getEmptyCellText = computed(() => {
 /** rowKey缓存 */
 const rowKeyGenStore = new WeakMap();
 
-const { onThDragStart, onThDragOver, onThDrop, isHeaderDraggable } = useThDrag({ props, emits });
+const { onThDragStart, onThDragOver, onThDrop, isHeaderDraggable } = useThDrag({ props, emits, colKeyGen });
 
 const { onTrDragStart, onTrDrop, onTrDragOver, onTrDragEnd, onTrDragEnter } = useTrDrag({ dataSourceCopy });